+++ src/test/java/com/ericsson/ei/extraction/test/TestExtractionHandler.java
@@ -31,7 +31,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import com.ericsson.ei.handlers.ExtractionHandler;
 import com.ericsson.ei.jmespath.JmesPathInterface;
 import com.ericsson.ei.rules.RulesObject;
-import com.ericsson.ei.rules.test.TestRulesHandler;
+import com.ericsson.ei.rules.TestRulesHandler;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
