+++ src/test/java/com/ericsson/ei/rules/TestRulesRestAPI.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.ericsson.ei.rules.test;
+package com.ericsson.ei.rules;
 
 import static org.junit.Assert.assertEquals;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
